refactor(train): split module staging out of registry.py into staging.py (#159)
registry.py was mixing three concerns: the registry (how to run modules),
the revision chain, and asset provisioning. Split:

- staging.py (new): ModuleStager + revision resolution - staging_revision()
  = REVISION (Colab Params form via WAKE_REVISION env, the default source
  whenever present) > baked wheel revision (hatch _revision.py) > main;
  repo_tarball_url(); _default_fetch.
- registry.py: back to the Registry concern only (Registry/RegistryError/
  _render). manager.py imports ModuleStager from .staging.
- Tests updated to the new homes; backend 53 passed. The notebook's
  os.environ['WAKE_REVISION'] = REVISION bridge is unchanged - the form
  value remains the default revision source.
