+++ apps/studio-backend/src/wake_training_service/registry.py
@@ -20,6 +20,10 @@ WAKE_<UPPER_KEY> per param - matching the notebook env convention
 (module.spec.json) remains the source of truth in the repo; the registry is
 the service's runtime view of it (a future step can generate it from specs).
 
+Asset provisioning for generic runtimes (Colab) lives in ``staging.py``
+(ModuleStager + revision resolution); ``registry.py`` stays "how to run a
+module".
+
 Entries may declare a ``stage`` spec (generic runtimes - e.g. Colab - have no
 repo checkout; the service stages the module's assets from the repo tarball
 on demand, driven by the job's moduleId - see ModuleStager).
@@ -30,50 +34,10 @@ from __future__ import annotations
 import json
 import os
 import re
-import tarfile
-import urllib.request
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any
 
 _TMPL = re.compile(r"\{([^}]+)\}")
-
-#: Repo tarball used by ModuleStager to stage module assets on demand.
-#: Fetch from the GitHub API tarball endpoint at the service's OWN build
-#: revision (baked into the wheel at build time, #159 option A) so the
-#: installed wheel and the staged module code can never drift; falls back to
-#: "main" when the revision metadata is absent (plain-directory builds).
-def _baked_revision() -> str:
-    try:
-        from wake_training_service import _revision  # type: ignore[attr-defined]
-
-        rev = getattr(_revision, "REVISION", "") or ""
-    except Exception:  # noqa: BLE001 - baked file missing (repo checkout / sdist)
-        rev = ""
-    return rev if rev else "main"
-
-
-def _staging_revision() -> str:
-    """Revision used for module staging: explicit override > baked wheel > main.
-
-    The Colab launcher notebook passes the Params-form ``REVISION`` value via
-    ``WAKE_REVISION`` (option A', #159) so the user's explicit choice wins
-    even if a stale wheel cached a different baked revision; self-hosted/CLI
-    runs without the env fall back to the wheel's baked revision (the two are
-    identical when the wheel was built from the pinned commit).
-    """
-    override = os.environ.get("WAKE_REVISION", "").strip()
-    if override:
-        return override
-    return _baked_revision()
-
-
-def repo_tarball_url(revision: str | None = None) -> str:
-    """GitHub API tarball endpoint - accepts a branch, tag, or commit SHA."""
-    return (
-        "https://api.github.com/repos/awareride/wake-studio/tarball/"
-        f"{revision or _staging_revision()}"
-    )
-
 
 class RegistryError(ValueError):
     pass
@@ -168,111 +132,6 @@ class Registry:
         for arg in entry.get("args", []):
             cmd.append(_render(arg, params, env))
         return cmd, str(cwd), env
-
-
-def _default_fetch(url: str, dest: Path) -> None:
-    """Stream-download a URL to ``dest`` (stdlib)."""
-    dest.parent.mkdir(parents=True, exist_ok=True)
-    req = urllib.request.Request(url, headers={"User-Agent": "wake-studio"})
-    with urllib.request.urlopen(req, timeout=120) as res, dest.open("wb") as out:
-        while True:
-            chunk = res.read(1 << 16)
-            if not chunk:
-                break
-            out.write(chunk)
-
-
-class ModuleStager:
-    """Stage registered-module assets on demand (generic runtimes, #159).
-
-    A Colab/generic runtime has no repo checkout, so the notebook stays
-    generic and the service provisions whatever module the job names: on the
-    first job for a module with a ``stage`` spec, the repo tarball is fetched
-    once and the module's declared paths are extracted under ``staged_root``.
-    Subsequent jobs reuse the staged tree (idempotent, no re-fetch).
-
-    Stage spec (per registry entry):
-
-        "stage": {
-          "paths": ["packages/modules/kws/streaming/train", "third_party/..."],
-          "cwd":   "packages/modules/kws/streaming/train",   // staged-root-relative
-          "env":   {"UPSTREAM_DIR": "third_party"}           // staged-root-relative
-        }
-    """
-
-    def __init__(
-        self,
-        staged_root: str | Path,
-        repo_url: str | None = None,
-        fetch: Callable[[str, Path], None] | None = None,
-    ) -> None:
-        self.staged_root = Path(staged_root)
-        self.repo_url = repo_url or repo_tarball_url()
-        self._fetch = fetch or _default_fetch
-        self._fetched = False
-
-    def prepare(
-        self, entry: dict[str, Any], base_dir: Path
-    ) -> tuple[Path, dict[str, str]] | None:
-        """-> (cwd, env overrides) if the entry stages assets, else None.
-
-        Returns None for entries without a stage spec (local repo checkout
-        layout - resolved against the registry base_dir as usual).
-        """
-        stage = entry.get("stage")
-        if not stage:
-            return None
-        script = entry.get("entry")
-        target = (self.staged_root / stage["cwd"]).resolve()
-        if (target / script).is_file():
-            return target, self._env_overrides(stage)
-        self._ensure_tarball()
-        self._extract(stage.get("paths", []))
-        if not (target / script).is_file():
-            raise RegistryError(
-                f"staged module '{entry.get('name', '?')}' missing entry "
-                f"{script} under {target}"
-            )
-        return target, self._env_overrides(stage)
-
-    def _env_overrides(self, stage: dict[str, Any]) -> dict[str, str]:
-        return {
-            k: str((self.staged_root / v).resolve())
-            for k, v in stage.get("env", {}).items()
-        }
-
-    def _ensure_tarball(self) -> None:
-        if self._fetched:
-            return
-        dest = self.staged_root / "_download" / "wake-studio.tar.gz"
-        dest.parent.mkdir(parents=True, exist_ok=True)
-        self._fetch(self.repo_url, dest)
-        self._fetched = True
-
-    def _extract(self, paths: list[str]) -> None:
-        """Extract repo-root-relative ``paths`` under staged_root (strip top dir)."""
-        archive = self.staged_root / "_download" / "wake-studio.tar.gz"
-        with tarfile.open(archive) as tf:
-            top = next(
-                (m.name.split("/", 1)[0] for m in tf.getmembers() if m.name.count("/")),
-                None,
-            )
-            if not top:
-                raise RegistryError(f"repo tarball {archive} is empty")
-            wanted = set()
-            for p in paths:
-                prefix = f"{top}/{p.strip('/')}"
-                wanted.update(
-                    m.name for m in tf.getmembers() if m.name == prefix or m.name.startswith(prefix + "/")
-                )
-            members = [m for m in tf.getmembers() if m.name in wanted]
-            for m in members:
-                m.name = m.name[len(top) + 1:]
-            try:
-                tf.extractall(self.staged_root, members=members, filter="data")
-            except TypeError:  # python < 3.12 (tarfile without filter kwarg)
-                tf.extractall(self.staged_root, members=members)
-            self.staged_root.mkdir(parents=True, exist_ok=True)
 
 
 def sys_executable() -> str:
